fix: /teams 관련 요청 정리

+++ back/src/mvp/team/teamService.js
@@ -13,27 +13,24 @@ import { folderService } from "../folder/folderService";
 import { bookmarkService } from "../bookmark/bookmarkService";
 
 class teamService {
-    static async createTeam({ manager, name, explanation }) {
+    static async createTeam({ manager_id, name, explanation }) {
         try {
             // 사용자 존재 확인
-            const user = await User.findOne({ where: { id: manager } });
-            if (!user) {
-                return getFailMsg({
-                    entity: "사용자",
-                    action: "조회",
-                });
+            const manager = await User.findOne({ where: { id: manager_id } });
+            if (!manager) {
+                return getFailMsg({ entity: "사용자", action: "조회" });
             }
 
             // 팀 생성
             const newTeam = await Team.create({
-                manager: user.id,
+                manager: manager.id,
                 name,
                 explanation,
             });
 
             // 팀의 첫번째 멤버(팀 매니저) 생성
             await Membership.create({
-                member_id: user.id,
+                member_id: manager.id,
                 team_id: newTeam.id,
             });
 
@@ -43,25 +40,22 @@ class teamService {
         }
     }
 
-    static async createMembership({ hostId, inviteeId, teamId }) {
+    static async createMembership({ host_id, invitee_id, team_id }) {
         try {
             // 초대한 사람, 초대받은 사람 존재 확인
-            const host = await User.findOne({ where: { id: hostId } });
+            const host = await User.findOne({ where: { id: host_id } });
             if (!host) {
                 return getFailMsg({ entity: "초대한 사용자", action: "조회" });
             }
-            const invitee = await User.findOne({
-                where: { id: inviteeId },
-            });
+            const invitee = await User.findOne({ where: { id: invitee_id } });
             if (!invitee) {
                 return getFailMsg({
                     entity: "초대받은 사용자",
                     action: "조회",
                 });
             }
-
-            // 멤버 추가 권한 확인
-            const team = await Team.findOne({ where: { id: teamId } });
+            // 멤버를 초대할 권한 확인
+            const team = await Team.findOne({ where: { id: team_id } });
             if (!team) {
                 return getFailMsg({ entity: "팀", action: "조회" });
             }
@@ -70,7 +64,6 @@ class teamService {
                     errorMessage: "현재 사용자는 팀에 초대할 권한이 없습니다.",
                 };
             }
-
             // 이미 회원인지 존재 확인
             const previousMembership = await Membership.count({
                 where: { member_id: invitee.id, team_id: team.id },
@@ -78,7 +71,6 @@ class teamService {
             if (previousMembership > 0) {
                 return { errorMessage: "이미 소속된 회원입니다." };
             }
-
             // 멤버 추가
             const membership = await Membership.create({
                 member_id: invitee.id,
@@ -94,10 +86,10 @@ class teamService {
     static async getTeamAll() {
         try {
             const teams = await sequelize.query(
-                `SELECT team.id as team_id, team.name, team.explanation, count(folder.id) as folder_count 
-            FROM ${Team.tableName} AS team 
-            LEFT JOIN ${Folder.tableName} as folder 
-            ON team.id = folder.team_id 
+                `SELECT team.id as team_id, team.name, team.explanation, count(folder.id) as folder_count
+            FROM ${Team.tableName} AS team
+                LEFT JOIN ${Folder.tableName} as folder
+                ON team.id = folder.team_id
             GROUP BY team.id;`,
                 { type: sequelize.QueryTypes.SELECT },
             );
@@ -133,63 +125,43 @@ class teamService {
         }
     }
 
-    static async getTeamsInfo({ ids }) {
-        try {
-            let result = await Team.findOne({
-                where: { id: ids },
-                include: [User],
-                raw: true,
-                nest: true,
-            });
+    // static async getTeamListUserJoined({ user_id }) {
+    //     try {
+    //         // 사용자 존재 확인
+    //         const user = await User.findOne({ where: { id: user_id } });
+    //         if (!user) {
+    //             return getFailMsg({ entity: "사용자", action: "조회" });
+    //         }
 
-            if (result.length === 0) {
-                return getFailMsg({ entity: "팀 상세정보", action: "조회" });
-            } else {
-                result.map(async (team) => {
-                    team["manager"] = team["user"];
-                    delete team["user"];
-                    await folderService.getFolderCntBookmarkCnt({ team_id });
-                    return team;
-                });
-            }
+    //         // 사용자가 속한 팀 조회
+    //         let teams = await sequelize.query(
+    //             `SELECT membership.team_id, team.name, team.explanation, count(folder.id) as folder_count
+    //             FROM (SELECT * FROM ${Membership.tableName} WHERE ${Membership.tableName}.member_id = ${user_id}) as membership
+    //             INNER JOIN ${Team.tableName} AS team
+    //             ON membership.team_id = team.id
+    //             LEFT JOIN ${Folder.tableName} as folder
+    //             ON team.id = folder.team_id
+    //             GROUP BY team.id;`,
+    //             { type: sequelize.QueryTypes.SELECT },
+    //         );
 
-            return result;
-        } catch (e) {
-            return { errorMessage: e };
-        }
-    }
-
-    static async getTeamListUserJoined({ user_id }) {
-        try {
-            // 사용자 존재 확인
-            const user = await User.findOne({ where: { id: user_id } });
-            if (!user) {
-                return getFailMsg({ entity: "사용자", action: "조회" });
-            }
-
-            // 사용자가 속한 팀 조회
-            let teams = await sequelize.query(
-                `SELECT membership.team_id, team.name, team.explanation, count(folder.id) as folder_count 
-                FROM (SELECT * FROM ${Membership.tableName} WHERE ${Membership.tableName}.member_id = ${user_id}) as membership
-                INNER JOIN ${Team.tableName} AS team 
-                ON membership.team_id = team.id 
-                LEFT JOIN ${Folder.tableName} as folder 
-                ON team.id = folder.team_id
-                GROUP BY team.id;`,
-                { type: sequelize.QueryTypes.SELECT },
-            );
-
-            return teams;
-        } catch (e) {
-            return { errorMessage: e };
-        }
-    }
+    //         return teams;
+    //     } catch (e) {
+    //         return { errorMessage: e };
+    //     }
+    // }
 
     static async getTeamMembers({ team_id }) {
         try {
+            // 팀 존재 확인
+            const team = await Team.findOne({ where: { id: team_id } });
+            if (!team) {
+                return getFailMsg({ entity: "팀", action: "조회" });
+            }
+            // 팀 멤버 조회
             const members = await sequelize.query(
-                `SELECT user.id, user.nickname, user.email, user.image_url, user.intro 
-                FROM (SELECT * FROM ${Membership.tableName} WHERE ${Membership.tableName}.team_id = ${team_id}) as membership 
+                `SELECT user.id, user.nickname, user.email, user.image_url, user.intro
+                FROM (SELECT * FROM ${Membership.tableName} WHERE ${Membership.tableName}.team_id = ${team_id}) as membership
                 INNER JOIN ${User.tableName} as user
                 ON membership.member_id = user.id`,
                 { type: sequelize.QueryTypes.SELECT },
@@ -248,16 +220,16 @@ class teamService {
         }
     }
 
-    static async updateTeamInfo({ teamId, requesterId, name, explanation }) {
+    static async updateTeamInfo({ team_id, requester_id, name, explanation }) {
         try {
             // 팀 아이디가 실제로 존재하는지
-            const team = await Team.findOne({ where: { id: teamId } });
+            const team = await Team.findOne({ where: { id: team_id } });
             if (!team) {
                 return getFailMsg({ entity: "팀", action: "조회" });
             }
             // 수정 요청자 requester가 실제로 존재하는지, 그리고 수정할 권한이 있는지
             const requester = await User.findOne({
-                where: { id: requesterId },
+                where: { id: requester_id },
             });
             if (!requester) {
                 return getFailMsg({ entity: "요청자", action: "조회" });
@@ -268,10 +240,10 @@ class teamService {
             //팀 상세정보 수정
             const [affectedRows] = await Team.update(
                 { name, explanation },
-                { where: { id: teamId, manager: requester.id } },
+                { where: { id: team.id, manager: requester.id } },
             );
             if (affectedRows === 0) {
-                return getFailMsg({ entity: "팀 상세정보", action: "수정" });
+                return { errorMessage: "서버에러" };
             }
             return getSuccessMsg({ entity: "팀 상세정보", action: "수정" });
         } catch (e) {
@@ -279,16 +251,16 @@ class teamService {
         }
     }
 
-    static async deleteMemeber({ teamId, memberId, requesterId }) {
+    static async deleteMemeber({ team_id, member_id, requester_id }) {
         try {
             // 팀 존재 확인
-            const team = await Team.findOne({ where: { id: teamId } });
+            const team = await Team.findOne({ where: { id: team_id } });
             if (!team) {
                 return getFailMsg({ entity: "팀", action: "조회" });
             }
             // 삭제 요청자 존재 확인 및 권한 확인
             const requester = await User.findOne({
-                where: { id: requesterId },
+                where: { id: requester_id },
             });
             if (!requester) {
                 return getFailMsg({ entity: "요청자", action: "조회" });
@@ -296,17 +268,23 @@ class teamService {
             if (team.manager !== requester.id) {
                 return { errorMessage: "사용자는 권한이 없습니다." };
             }
-            // 사용자 존재 확인
-            const member = await User.findOne({ where: { id: memberId } });
+            // 사용자 존재 확인 및 소속 여부
+            const member = await User.findOne({ where: { id: member_id } });
             if (!member) {
                 return getFailMsg({ entity: "사용자", action: "조회" });
             }
+            const membership = await Membership.findOne({
+                where: { member_id: member.id, team_id: team.id },
+            });
+            if (!membership) {
+                return getFailMsg({ entity: "멤버십", action: "조회" });
+            }
             // 회원 삭제
             const result = await Membership.destroy({
-                where: { team_id: teamId, member_id: memberId },
+                where: { team_id: team.id, member_id: member.id },
             });
             if (result === 0) {
-                return getFailMsg({ entity: "회원", action: "삭제" });
+                return { errorMessage: "서버에러" };
             }
             return getSuccessMsg({ entity: "회원", action: "삭제" });
         } catch (e) {
@@ -314,16 +292,16 @@ class teamService {
         }
     }
 
-    static async deleteTeam({ teamId, requesterId }) {
+    static async deleteTeam({ team_id, requester_id }) {
         try {
             // 팀 존재 확인
-            const team = await Team.findOne({ where: { id: teamId } });
+            const team = await Team.findOne({ where: { id: team_id } });
             if (!team) {
                 return getFailMsg({ entity: "팀", action: "조회" });
             }
             // 삭제 요청자 존재 확인 및 권한 확인
             const requester = await User.findOne({
-                where: { id: requesterId },
+                where: { id: requester_id },
             });
             if (!requester) {
                 return getFailMsg({ entity: "요청자", action: "조회" });
@@ -333,10 +311,10 @@ class teamService {
             }
             // 팀 삭제
             const result = await Team.destroy({
-                where: { id: teamId },
+                where: { id: team.id },
             });
             if (result === 0) {
-                return getFailMsg({ entity: "회원", action: "삭제" });
+                return { errorMessage: "서버에러" };
             }
             return getSuccessMsg({ entity: "회원", action: "삭제" });
         } catch (e) {
